requirements.txt
mysql-connector-python==8.0.32
requests==2.28.2
python-dotenv~=0.21.1
beautifulsoup4~=4.11.2
colorama~=0.4.6